+++ src/lib.ts
@@ -1,7 +1,7 @@
 export { Decimal } from 'decimal.js';
 export { CharString } from './CharString';
 export { ComplexDecimal } from './ComplexDecimal';
-export { MultiArray } from './MultiArray';
+export { ElementType, MultiArray } from './MultiArray';
 export { Structure } from './Structure';
 export { FunctionHandle } from './FunctionHandle';
 export { LinearAlgebra } from './LinearAlgebra';